Patch for bug #129: Removing Events out of activities * All setup/connect methods only take an EventProcessor and no   longer a ThreadInterface or ActivityInterface * The CompletionProcessor is both an EventProcessor and a   NonPeriodicActivity
In order to emulate old behaviour, inherit from EventProcessor or
BlockingEventProcessor or add such an object as a member to your class.
Then pass the appropriate pointer to the Event setup/connect methods.
Finally invoke on your EventProcessor step() to process a single event 
or loop() to wait for events.



git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/orocos-trunk@5276 ce417995-dfc9-0310-95a0-acdaff106893