rtt: implement event-driven schemes
This commit introduces a data-driven scheme in which a task is woken up
when new data is available on one of its input ports. To achieve this,
the following modifications have been done:
 * a per-input-port Event< void() > object can be created, which is
   signalled every time new data is written on this port.
 * the EventDrivenActivity has been resurrected (the current one could
   not even compile because of pure-virtual methods not being
   implemented). The new implementation is based on NonPeriodicActivity.
   It provides an addEvent method which defines which events should
   trigger the activity.
 * a DataDrivenTask, child of TaskContext, is provided. This glues
   everything by calling the addEvent method of its activity (which has
   to be an EventDrivenActivity) for all the trigger events of its input
   ports.

This is still a partial implementation as the port events are signalled
only if an input port has written on the connection. It won't work if
the connection object is used directly.

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@29744 ce417995-dfc9-0310-95a0-acdaff106893