event-driven: make DataDrivenTask and EventDrivenActivity RT-friendly
The current implementation was using std::set to filter duplicates in
the wakeup events. The problem is that std::set does allocate memory
behind our backs, making it unusable in RT contexts.

This patch changes the container from std::set to std::vector and, given
that we know the maximum number of unique events to wake us up,
reserves() that number on startup. Then, duplicates are searched "by
hand". It could have a performance impact on tasks that listen to a high
number of events, but that seems an unlikely situation. One would have
to not only listen to a high number of events, but the event-driven
activity would have to be scheduled so that most of these events fire
before the activity gets fired.

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@29748 ce417995-dfc9-0310-95a0-acdaff106893