Completes the default SequentialActivity setup for TaskContext
With this patch, any TaskContext is created with a 'sequential activity' by
default. This means it can be started (or activated) and process commands and
events, without having a thread. As a bonus, thread-safety is still fully
guaranteed for these primitives and the behaviour is very similar (if not equal) to
NonPeriodicActivity. When a command or event comes in, it is processed in the
thread of the sender (ie sequential with the calling code). If multiple
senders operate on the receiver, the first one processes all the requests, so
no thread blocks in case of contention in this implementation.

Please review and test this.

New API:

TaskContext a(A);
a.start(); // works
a.stop();
a.setActivity( new PeriodicActivity(0, 0.01) ); // assign new activity to 'A'
a.start(); now 'A' is periodic.
a.stop();
a.setActivity( 0 ); // make 'A' sequential again.
// etc...

This patch is backwards compatible with 'old' applications.

From: Peter Soetens <peter.soetens@fmtc.be>

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@29905 ce417995-dfc9-0310-95a0-acdaff106893