corba: cleanup 0-values returned through Corba
This patch fixes a problem and introduces a "best practice". NULL is not
a valid value for _ptr types in Corba. "Invalid pointer" must be created
through the calls to the typed _nil() methods.

Moreover, it is better to return "false" for booleans, as it improves
the ability to search for bad "return 0" cases.

From: Sylvain Joyeux <sylvain.joyeux@dfki.de>

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@29908 ce417995-dfc9-0310-95a0-acdaff106893