corba: actually get a ControlTaskProxy object in corba tests
ControlTaskProxy::Create (rightly) returns a local object when a local
servant is given to it. The problem is that, in corba tests, we actually
want to use a corba proxy for the tests to be valid.

This commit adds a force_remote parameter to ControlTaskProxy::Create,
so that it creates a local proxy even in the case where the ControlTask
pointer is actually a local task.

From: Sylvain Joyeux <sylvain.joyeux@dfki.de>

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@29914 ce417995-dfc9-0310-95a0-acdaff106893