corba: fix an ambiguity problem in corba-to-C++ convertions
When:
  * using OmniORB
  * on an x32 machine

it seems that there is an ambiguity problem for int-to-corba
and uint-to-corba convertions. I guess it has to choose between
int-to-Long and uint-to-ULong or something.

Given that int is always 32bits on both architectures, we make the compiler
chose  Long/ULong.

From: Daniel Kühn <dakuehn@kuehn2-u.dfki.uni-bremen.de>

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@29918 ce417995-dfc9-0310-95a0-acdaff106893