Applied initial ARM port for RTT.
This port was tested on a gumstix and contributed by
Felipe Brandão Cavalcanti <cavalkaf@gmail.com>:

Regarding the ARM version of OROCOS, the version that I wrote seems to be
working fine - the entire system and its dependencies compile and run
without any issues. However, there is still the issue of the IRQs that I
will leave up to you guys - I really don't have enough experience in this
field to deal with such matters.

I am only sending in the files, as I am not sure on how to send in a proper
patch...

I wrote/tested this patch with OROCOS 1.6.1 (as well as 1.6.0), Xenomai
2.4.6.1 and Linux 2.6.24 on an PXA-270 ARM platform running OpenEmbedded
(the Gumstix Verdex).