Various portable patches to compile the code on MSVC2005
The only fuzzy part about this patch is the typeof replacement
MSVC2005 doesn't support it and some hand-coded constructs needed to
be put in place. But it works on gnu as well.

This patch was slightly adapted and trimmed from the original
patch of Simon.

From: Simon Pelletier-Thibault <pelletierts@amdinc.ca>

git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/trunk/rtt@30319 ce417995-dfc9-0310-95a0-acdaff106893